Today I created the footer and a button
The footer contains some information about the company, and the button can be used to navigate between sections.
Note: When we click on "Cardápio" (in English "menu") it directs the user to a section where we can see some types of food, but it's not the menu yet. I'll try to make another page to put the menu and when the button is pressed it will open in a new tab.